Query all XNAT scan resources

Query all scan resources to match specific query parameters and return a list of all matching rows

query_scan_resources(conn, ...)

Arguments

conn
The XNAT connection returned by a xnat_connect call

...
Select query parameters: subject_ID, project, age, experiment_ID, type TR, TE, TI, flip, voxel_res, voxel_res_X, voxel_res_Y, voxel_res_Z, orientation

Value

A data.frame containing all matching rows. XNAT does not do sql join joins so only one row is returned per match. However each of the experiment_IDs returned will have at least one row matching the user query (even if the displayed results show something else). This function should be used just to retrieve matching experiment IDs for downloading the queried data.

Examples

hcp <-xnat_connect('https://db.humanconnectome.org', xnat_name = "hcp")

query_scan_resources(hcp,age='26', project='HCP_500')

#> Error in xnat_call("/data/search?format=csv", data = get_scan_parameters_search_xml(...)): error during HTTP request